[ML] Refactor autodetect service into its own class
This also improves aims to improve the corresponding unit tests
with regard to readability and maintainability.
